Added basic extension structure, can read grid size
Created the extension (including icons), and did some basic DOM magic
to determine the (most likely) size of the puzzle grid.

Current solution relies on guessing/calculating it out of the size of the
board-containing element, which could break if the sizing/spacing changes.
The alternative was to look in the "puzzleInfo"-class element and parse
text to get the size, but for that we'd need to hard-code the challenge
puzzle sizes, since those only say e.g. "Daily Challenge".
Note that, for the sake of fairness, we might want to prevent users from
applying the extension to daily puzzles anyway?

Currently the extension pop-up (when clicked in the Extensions toolbar)
is just a stub, too.