Small edits & Renames
Renamed 'NubCell' to 'BulbCell' (the cell type that has only one outgoing
connection).

Added tracking for whether a cell is pinned and for keeping links to its
neighbouring cells.

Removed support for cells remembering their blocked (cannot have a link
that way) and forced (must have a link that way) connections. This change
might get reverted if a "solve all" solution is implemented, but it
increasingly seems like interacting with the page via code won't work,
so that might not happen after all.