Added basic solving functionality to cell classes.
Currently, this is implemented via each cell having an array of strategies,
and iterating through all of them to try to find a correct solution.
I'm not sure I like the placement of that code, though - perhaps the
strategies should go into their own file or maybe not even be members of
the cell classes?