Bugfixes + changed what attemptSolve() returns.
Changed attemptSolve() so it no longer modifies the cell, but instead
returns a non-DOM-attached cell aligned as the correct alignment should be.
Basically, it says "here's how I should look in the solved position".
This is a bit inefficient since we create an entire cell object; it might
be better to just return a cell type and correct facing direction, and
then write some code to display that nicely instead of leaning on toString.

Also changed the rotate() and togglePinned() methods to no longer attempt
to send HTML events, since the webpage does not accept them (and sending
them was messing with our own event listeners that keep in sync with the
grid).

Bugfixes:
- fixed ElbowCell indicating the wrong connection directions when pinned
- fixed a typo in ThreeWayCell strategy ThreeDirectionsConnected