Changed "solve one" functionality to improve user experience
Instead of simply displaying a message like "rotate the highlighted cell
until it looks like this", the highlight is now colour-coded and the user
simply needs to click until the cell is green. There are also listeners
to keep track of user actions and update/remove this highlight.

With this change, AbstractCell's attemptSolve() method was changed (again)
to simply return the correct facing direction for the cell. The new
Direction helper functions are then used to calculate how many rotations
away the cell is from being correct (passing through a cell method, so
LineCell can account for only needing one rotation at most to be correct).

Added new CSS classes for the new highlight colours to pipesSolver.css.

Also fixed a minor bug in the cell DOM disteners, where the cell would
register a rotation even if it was pinned when it got clicked (and hence
didn't actually rotate).