Improved "solve one" functionality
Made the "solve one" button skip cells that are already highlighted for
the user, so that every button press highlights a new cell if the solver
is capable of that. (Previously, an already-highlighted cell could be
"hit" again if the user hadn't pinned it yet, which felt odd - you'd press
the button and nothing would happen.)

Also made the solver pick a starting point at random when picking a cell
to solve, so that highlighted cells feel more spontaneous instead of
always going top-down left-to-right.